refactor: TypeUrl as constants
