+++ packages/chain-sdk/src/constants/index.ts
@@ -1,2 +1,4 @@
+export * from './typeUrl';
+
 export const ZERO_PUBKEY = '0x000000000000000000000000000000000000000000000000000000000000000000';
-export const DEFAULT_DENOM = ' BNB';
+export const DEFAULT_DENOM = 'BNB';
