+++ packages/chain-sdk/src/constants/typeUrl.ts
@@ -1,0 +1,29 @@
+export const MsgSendTypeUrl = '/cosmos.bank.v1beta1.MsgSend';
+export const MsgMultiSendTypeUrl = '/cosmos.bank.v1beta1.MsgMultiSend';
+export const MsgClaimTypeUrl = '/cosmos.oracle.v1.MsgClaim';
+export const MsgGrantAllowanceTypeUrl = '/cosmos.feegrant.v1beta1.MsgGrantAllowance';
+export const MsgRevokeAllowanceTypeUrl = '/cosmos.feegrant.v1beta1.MsgRevokeAllowance';
+export const MsgTransferOutTypeUrl = '/greenfield.bridge.MsgTransferOut';
+export const MsgAttestTypeUrl = '/greenfield.challenge.MsgAttest';
+export const MsgSubmitTypeUrl = '/greenfield.challenge.MsgSubmit';
+export const MsgCreatePaymentAccountTypeUrl = '/greenfield.payment.MsgCreatePaymentAccount';
+export const MsgDepositTypeUrl = '/greenfield.payment.MsgDeposit';
+export const MsgDisableRefundTypeUrl = '/greenfield.payment.MsgDisableRefund';
+export const MsgWithdrawTypeUrl = '/greenfield.payment.MsgWithdraw';
+export const MsgCancelCreateObjectTypeUrl = '/greenfield.storage.MsgCancelCreateObject';
+export const MsgCreateBucketTypeUrl = '/greenfield.storage.MsgCreateBucket';
+export const MsgCreateGroupTypeUrl = '/greenfield.storage.MsgCreateGroup';
+export const MsgCreateObjectTypeUrl = '/greenfield.storage.MsgCreateObject';
+export const MsgDeleteBucketTypeUrl = '/greenfield.storage.MsgDeleteBucket';
+export const MsgDeleteGroupTypeUrl = '/greenfield.storage.MsgDeleteGroup';
+export const MsgDeleteObjectTypeUrl = '/greenfield.storage.MsgDeleteObject';
+export const MsgDeletePolicyTypeUrl = '/greenfield.storage.MsgDeletePolicy';
+export const MsgLeaveGroupTypeUrl = '/greenfield.storage.MsgLeaveGroup';
+export const MsgMirrorBucketTypeUrl = '/greenfield.storage.MsgMirrorBucket';
+export const MsgMirrorGroupTypeUrl = '/greenfield.storage.MsgMirrorGroup';
+export const MsgMirrorObjectTypeUrl = '/greenfield.storage.MsgMirrorObject';
+export const MsgPutPolicyTypeUrl = '/greenfield.storage.MsgPutPolicy';
+export const MsgUpdateBucketInfoTypeUrl = '/greenfield.storage.MsgUpdateBucketInfo';
+export const MsgUpdateGroupExtraTypeUrl = '/greenfield.storage.MsgUpdateGroupExtra';
+export const MsgUpdateGroupMemberTypeUrl = '/greenfield.storage.MsgUpdateGroupMember';
+export const MsgUpdateObjectInfoTypeUrl = '/greenfield.storage.MsgUpdateObjectInfo';
