+++ packages/chain-sdk/src/messages/bank/MsgSend.ts
@@ -1,4 +1,3 @@
-export const MsgSendTypeUrl = '/cosmos.bank.v1beta1.MsgSend';
 export const MsgSendSDKTypeEIP712 = {
   Msg1: [
     {
