+++ packages/chain-sdk/src/messages/feegrant/MsgRevokeAllowance.ts
@@ -1,4 +1,3 @@
-export const MsgRevokeAllowanceTypeUrl = '/cosmos.feegrant.v1beta1.MsgRevokeAllowance';
 export const MsgRevokeAllowanceSDKTypeEIP712 = {
   Msg1: [
     {
