+++ packages/chain-sdk/src/messages/greenfield/chanenge/MsgAttest.ts
@@ -1,5 +1,3 @@
-export const MsgAttestTypeUrl = '/greenfield.challenge.MsgAttest';
-
 export const MsgAttestSDKTypeEIP712 = {
   Msg1: [
     {
