+++ packages/chain-sdk/src/messages/greenfield/chanenge/MsgSubmit.ts
@@ -1,5 +1,3 @@
-export const MsgSubmitTypeUrl = '/greenfield.challenge.MsgSubmit';
-
 export const MsgSubmitSDKTypeEIP712 = {
   Msg1: [
     {
