+++ packages/chain-sdk/src/messages/greenfield/payment/MsgCreatePaymentAccount.ts
@@ -1,5 +1,3 @@
-export const MsgCreatePaymentAccountTypeUrl = '/greenfield.payment';
-
 export const MsgCreatePaymentAccountSDKTypeEIP712 = {
   Msg: [
     {
