+++ packages/chain-sdk/src/messages/greenfield/payment/MsgDeposit.ts
@@ -1,4 +1,3 @@
-export const MsgDepositTypeUrl = '/greenfield.payment.MsgDeposit';
 export const MsgDepositSDKTypeEIP712 = {
   Msg1: [
     {
