+++ packages/chain-sdk/src/messages/greenfield/storage/MsgCreateGroup.ts
@@ -1,5 +1,3 @@
-export const MsgCreateGroupTypeUrl = '/greenfield.storage.MsgCreateGroup';
-
 export const MsgCreateGroupSDKTypeEIP712 = {
   Msg1: [
     {
