+++ packages/chain-sdk/src/messages/greenfield/storage/MsgCreateObject.ts
@@ -1,5 +1,3 @@
-export const MsgCreateObjectTypeUrl = '/greenfield.storage.MsgCreateObject';
-
 export const MsgCreateObjectSDKTypeEIP712 = {
   Msg1: [
     {
