+++ packages/chain-sdk/src/messages/greenfield/storage/MsgDeleteBucket.ts
@@ -1,5 +1,3 @@
-export const MsgDeleteBucketTypeUrl = '/greenfield.storage.MsgDeleteBucket';
-
 export const MsgDeleteBucketSDKTypeEIP712 = {
   Msg1: [
     {
