+++ packages/chain-sdk/src/messages/greenfield/storage/MsgDeleteGroup.ts
@@ -1,4 +1,3 @@
-export const MsgDeleteGroupTypeUrl = '/greenfield.storage.MsgDeleteGroup';
 export const MsgDeleteGroupSDKTypeEIP712 = {
   Msg1: [
     {
