+++ packages/chain-sdk/src/messages/greenfield/storage/MsgDeletePolicy.ts
@@ -1,5 +1,3 @@
-export const MsgDeletePolicyTypeUrl = '/greenfield.storage.MsgDeletePolicy';
-
 export const MsgDeletePolicySDKTypeEIP712 = {
   Msg1: [
     {
