+++ packages/chain-sdk/src/messages/greenfield/storage/MsgLeaveGroup.ts
@@ -1,4 +1,3 @@
-export const MsgLeaveGroupTypeUrl = '/greenfield.storage.MsgLeaveGroup';
 export const MsgLeaveGroupSDKTypeEIP712 = {
   Msg1: [
     {
