+++ packages/chain-sdk/src/messages/greenfield/storage/MsgMirrorBucket.ts
@@ -1,5 +1,3 @@
-export const MsgMirrorBucketTypeUrl = '/greenfield.storage.MsgMirrorBucket';
-
 export const MsgMirrorBucketSDKTypeEIP712 = {
   Msg1: [
     {
