+++ packages/chain-sdk/src/messages/greenfield/storage/MsgMirrorObject.ts
@@ -1,5 +1,3 @@
-export const MsgMirrorObjectTypeUrl = '/greenfield.storage.MsgMirrorObject';
-
 export const MsgMirrorObjectSDKTypeEIP712 = {
   Msg1: [
     {
