+++ packages/chain-sdk/src/messages/greenfield/storage/MsgUpdateGroupMember.ts
@@ -1,5 +1,3 @@
-export const MsgUpdateGroupMemberTypeUrl = '/greenfield.storage.MsgUpdateGroupMember';
-
 export const MsgUpdateGroupMemberSDKTypeEIP712 = {
   Msg1: [
     {
