+++ packages/chain-sdk/src/messages/greenfield/storage/MsgUpdateObjectInfo.ts
@@ -1,5 +1,3 @@
-export const MsgUpdateObjectInfoTypeUrl = '/greenfield.storage.MsgUpdateObjectInfo';
-
 export const MsgUpdateObjectInfoSDKTypeEIP712 = {
   Msg1: [
     {
