--- packages/chain-sdk/src/tx/index.ts
@@ -1,4 +1,0 @@
-export interface IRawTxInfo {
-  bytes: Uint8Array;
-  hex: string;
-}
